Setting up the backend
We initialize the backend in the root folder. The reason we do this is becuase when we deploy the app we want to start it from the root folder.

We install express and nodemon. In package.json file we add the scripts for running our app in development mode and production mode. Development script ===> "run": "nodemon server/index.js" and production script ===> "start": "node server/index.js"

We create index.js file in the server folder and write a basic boilerplate code to start the server. We also move the .gitignore file from client folder to the root folder.