Set up mongodb databse
We install dotenv to manage environment variables.

We setup a mongodb database on the mongodb atlas. We created a .env file that contains the PORT number and the MONGODB_URI. The .env is included in the root directory of the project. Add this .env file to .gitignore.

We create a new file called db.js in the server folder and write a function to connect to our database. Then in our index.js file we call that function and use .then .catch to make sure our server only starts if there the connection to database is established.