Changing database connection configuration
Previously we start our server by using .then and .catch on the connectDB function to make sure that the server only starts when the connection to database is established. It is not working somehow. SO we removed .then and .catch on the connectDB function and just call it normally to established the connection. Then use app.listen() to start the server.