Created a utils folder and added file for custom error handling
Sometimes we need to create and send custom errors to the user. These errors don't occur in the server but we want to send a error to the user anyways. Suppose a user forgets to fill some data in the form, then we need to send him an error telling him that they missed a field in the form.